Add files via upload
Forgot to upload and update for like a week. Articles is a new workplace for animations and shape creation. Clover is not updated but is an option in the header top. Articles and index are in a loop now. clover is a one way. index is cleaned up for now. I'll minimize code as need.